+++ facade.go
@@ -3,6 +3,7 @@ package minibus
 import (
 	"context"
 	"reflect"
+	"runtime/trace"
 )
 
 // Run exchanges messages between functions that it executes in parallel.
@@ -11,10 +12,12 @@ import (
 // error, or ctx is canceled. Functions are added using the [WithFunc] option.
 func Run(ctx context.Context, options ...Option) error {
 	s := &session{
-		inboxSize: 10,
-		ready:     make(chan *function),
-		returned:  make(chan *function),
-		bus:       make(chan envelope),
+		InboxSize:         10,
+		Ready:             make(chan *function),
+		Returned:          make(chan *function),
+		Bus:               make(chan envelope),
+		funcs:             map[*function]struct{}{},
+		subscribersByType: map[messageType]*subscribers{},
 	}
 
 	for _, applyOption := range options {
@@ -22,8 +25,8 @@ func Run(ctx context.Context, options ...Option) error {
 	}
 
 	// Only create inboxes after all [WithInboxSize] options have been applied.
-	for f := range s.funcs.Elements() {
-		f.inbox = make(chan any, s.inboxSize)
+	for f := range s.funcs {
+		f.Inbox = make(chan any, s.InboxSize)
 	}
 
 	return s.run(ctx)
@@ -35,12 +38,13 @@ type Option func(*session)
 // WithFunc adds a function to be executed by a call to [Run].
 func WithFunc(fn func(context.Context) error) Option {
 	return func(s *session) {
-		s.funcs.Add(&function{
-			impl:     fn,
-			session:  s,
-			outbox:   make(chan any),
-			returned: make(chan struct{}),
-		})
+		s.funcs[&function{
+			Session:       s,
+			Subscriptions: map[messageType]struct{}{},
+			Outbox:        make(chan any),
+			Returned:      make(chan struct{}),
+			impl:          fn,
+		}] = struct{}{}
 	}
 }
 
@@ -48,7 +52,7 @@ func WithFunc(fn func(context.Context) error) Option {
 // buffered in each function's inbox.
 func WithInboxSize(size int) Option {
 	return func(s *session) {
-		s.inboxSize = size
+		s.InboxSize = size
 	}
 }
 
@@ -59,12 +63,18 @@ func WithInboxSize(size int) Option {
 // must be called before [Ready].
 func Subscribe[M any](ctx context.Context) {
 	f := caller(ctx)
-	if f.ready {
+	if f.Ready {
 		panic("minibus: Subscribe() must not be called after calling Ready()")
 	}
 
-	t := reflect.TypeFor[M]()
-	f.subscriptions.Add(t)
+	t := messageType{reflect.TypeFor[M]()}
+	f.Subscriptions[t] = struct{}{}
+
+	if t.Kind() == reflect.Interface {
+		trace.Logf(ctx, "minibus", "subscribed to messages that implement %q", t)
+	} else {
+		trace.Logf(ctx, "minibus", "subscribed to %q messages", t)
+	}
 }
 
 // Ready signals that the function has made all relevant [Subscribe] calls and
@@ -74,16 +84,18 @@ func Subscribe[M any](ctx context.Context) {
 // [Run] have called [Ready].
 func Ready(ctx context.Context) {
 	f := caller(ctx)
-	if f.ready {
+	if f.Ready {
 		return
 	}
 
+	trace.Logf(ctx, "minibus", "ready to exchange messages, %d subscription(s)", len(f.Subscriptions))
+
 	select {
 	case <-ctx.Done():
-	case f.session.ready <- f:
+	case f.Session.Ready <- f:
 	}
 
-	f.ready = true
+	f.Ready = true
 }
 
 // Inbox returns the channel on which the function receives messages send by
@@ -95,7 +107,7 @@ func Ready(ctx context.Context) {
 // No messages are delivered until all functions executed by the same call to
 // [Run] have called [Ready].
 func Inbox(ctx context.Context) <-chan any {
-	return caller(ctx).inbox
+	return caller(ctx).Inbox
 }
 
 // Outbox returns a channel on which the function can send messages to other
@@ -104,7 +116,7 @@ func Inbox(ctx context.Context) <-chan any {
 // The channel will block until all functions executed by the same call to [Run]
 // have called [Ready].
 func Outbox(ctx context.Context) chan<- any {
-	return caller(ctx).outbox
+	return caller(ctx).Outbox
 }
 
 // Send sends a message, or returns an error if ctx is canceled.