+++ CHANGELOG.md
@@ -10,12 +10,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.2.1] - 2024-08-01
 
 ### Added
 
 - `Subscribe()` now supports subscribing to interface types. The function will
   receive any messages that implement the interface.
+- Added `runtime/trace` task and log annotations.
 
 ## [0.2.0] - 2024-07-31
 
@@ -47,6 +48,7 @@ functions executed by `Run()` as "functions".
 [0.1.0]: https://github.com/dogmatiq/minibus/releases/tag/v0.1.0
 [0.1.1]: https://github.com/dogmatiq/minibus/releases/tag/v0.1.1
 [0.2.0]: https://github.com/dogmatiq/minibus/releases/tag/v0.2.0
+[0.2.1]: https://github.com/dogmatiq/minibus/releases/tag/v0.2.1
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD